Refactor: split as TaskTabs component
